create pytest_indexd utils and move functions
