+++ tests/integration/test_client.py
@@ -5,7 +5,7 @@ import pytest
 from requests import HTTPError
 
 if os.getenv("USE_PYTEST_INDEXD", "false").lower() == "true":
-    from pytest_indexd import create_random_index, create_random_index_version
+    from pytest_indexd.utils import create_random_index, create_random_index_version
 else:
     from indexd_test_utils import create_random_index, create_random_index_version
